+++ realm/realm-library/src/main/cpp/io_realm_internal_Table.cpp
@@ -20,8 +20,19 @@
 #include "io_realm_internal_Table.h"
 #include "tablebase_tpl.hpp"
 
+#include "util/format.hpp"
+
+#include "jni_util/java_exception_thrower.hpp"
+#include "java_exception_def.hpp"
+
 using namespace std;
 using namespace realm;
+using namespace realm::_impl;
+using namespace realm::jni_util;
+using namespace realm::util;
+
+static const char* c_null_values_cannot_set_required_msg = "The primary key field '%1' has 'null' values stored.  It "
+                                                           "cannot be converted to a '@Required' primary key field.";
 
 static void finalize_table(jlong ptr);
 
@@ -192,11 +203,12 @@ JNIEXPORT jboolean JNICALL Java_io_realm_internal_Table_nativeIsColumnNullable(J
 // 6. removing the original column and renaming the temporary column will make it look like original is being modified
 
 JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNullable(JNIEnv* env, jobject,
-                                                                                  jlong nativeTablePtr,
-                                                                                  jlong columnIndex)
+                                                                                  jlong native_table_ptr,
+                                                                                  jlong j_column_index,
+                                                                                  jboolean is_primary_key)
 {
-    Table* table = TBL(nativeTablePtr);
-    if (!TBL_AND_COL_INDEX_VALID(env, table, columnIndex)) {
+    Table* table = TBL(native_table_ptr);
+    if (!TBL_AND_COL_INDEX_VALID(env, table, j_column_index)) {
         return;
     }
     if (table->has_shared_type()) {
@@ -204,7 +216,7 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNullabl
         return;
     }
     try {
-        size_t column_index = S(columnIndex);
+        size_t column_index = S(j_column_index);
         if (table->is_nullable(column_index)) {
             return; // column is already nullable
         }
@@ -232,12 +244,22 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNullabl
             j++;
         }
 
+        // Search index has too be added first since if it is a PK field, add_xxx_unique will check it.
+        if (table->has_search_index(column_index + 1)) {
+            table->add_search_index(column_index);
+        }
+
         for (size_t i = 0; i < table->size(); ++i) {
             switch (column_type) {
                 case type_String: {
                     // Payload copy is needed
                     StringData sd(table->get_string(column_index + 1, i));
-                    table->set_string(column_index, i, sd);
+                    if (is_primary_key) {
+                        table->set_string_unique(column_index, i, sd);
+                    }
+                    else {
+                        table->set_string(column_index, i, sd);
+                    }
                     break;
                 }
                 case type_Binary: {
@@ -248,7 +270,12 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNullabl
                     break;
                 }
                 case type_Int:
-                    table->set_int(column_index, i, table->get_int(column_index + 1, i));
+                    if (is_primary_key) {
+                        table->set_int_unique(column_index, i, table->get_int(column_index + 1, i));
+                    }
+                    else {
+                        table->set_int(column_index, i, table->get_int(column_index + 1, i));
+                    }
                     break;
                 case type_Bool:
                     table->set_bool(column_index, i, table->get_bool(column_index + 1, i));
@@ -273,9 +300,6 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNullabl
                     return;
             }
         }
-        if (table->has_search_index(column_index + 1)) {
-            table->add_search_index(column_index);
-        }
         table->remove_column(column_index + 1);
         table->rename_column(table->get_column_index(tmp_column_name), column_name);
     }
@@ -283,11 +307,12 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNullabl
 }
 
 JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNotNullable(JNIEnv* env, jobject,
-                                                                                     jlong nativeTablePtr,
-                                                                                     jlong columnIndex)
+                                                                                     jlong native_table_ptr,
+                                                                                     jlong j_column_index,
+                                                                                     jboolean is_primary_key)
 {
-    Table* table = TBL(nativeTablePtr);
-    if (!TBL_AND_COL_INDEX_VALID(env, table, columnIndex)) {
+    Table* table = TBL(native_table_ptr);
+    if (!TBL_AND_COL_INDEX_VALID(env, table, j_column_index)) {
         return;
     }
     if (table->has_shared_type()) {
@@ -295,7 +320,7 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNotNull
         return;
     }
     try {
-        size_t column_index = S(columnIndex);
+        size_t column_index = S(j_column_index);
         if (!table->is_nullable(column_index)) {
             return; // column is already not nullable
         }
@@ -322,23 +347,39 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNotNull
             j++;
         }
 
+        // Search index has too be added first since if it is a PK field, add_xxx_unique will check it.
+        if (table->has_search_index(column_index + 1)) {
+            table->add_search_index(column_index);
+        }
+
         for (size_t i = 0; i < table->size(); ++i) {
             switch (column_type) { // FIXME: respect user-specified default values
                 case type_String: {
                     StringData sd = table->get_string(column_index + 1, i);
                     if (sd == realm::null()) {
-                        table->set_string(column_index, i, "");
+                        if (is_primary_key) {
+                            THROW_JAVA_EXCEPTION(env, JavaExceptionDef::IllegalState,
+                                                 format(c_null_values_cannot_set_required_msg, column_name));
+                        }
+                        else {
+                            table->set_string(column_index, i, "");
+                        }
                     }
                     else {
                         // Payload copy is needed
-                        table->set_string(column_index, i, sd);
+                        if (is_primary_key) {
+                            table->set_string_unique(column_index, i, sd);
+                        }
+                        else {
+                            table->set_string(column_index, i, sd);
+                        }
                     }
                     break;
                 }
                 case type_Binary: {
                     BinaryData bd = table->get_binary(column_index + 1, i);
                     if (bd.is_null()) {
-                        table->set_binary(column_index, i, BinaryData(""));
+                        table->set_binary(column_index, i, BinaryData("", 0));
                     }
                     else {
                         // Payload copy is needed
@@ -349,10 +390,21 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNotNull
                 }
                 case type_Int:
                     if (table->is_null(column_index + 1, i)) {
-                        table->set_int(column_index, i, 0);
+                        if (is_primary_key) {
+                            THROW_JAVA_EXCEPTION(env, JavaExceptionDef::IllegalState,
+                                                 format(c_null_values_cannot_set_required_msg, column_name));
+                        }
+                        else {
+                            table->set_int(column_index, i, 0);
+                        }
                     }
                     else {
-                        table->set_int(column_index, i, table->get_int(column_index + 1, i));
+                        if (is_primary_key) {
+                            table->set_int_unique(column_index, i, table->get_int(column_index + 1, i));
+                        }
+                        else {
+                            table->set_int(column_index, i, table->get_int(column_index + 1, i));
+                        }
                     }
                     break;
                 case type_Bool:
@@ -398,9 +450,6 @@ JNIEXPORT void JNICALL Java_io_realm_internal_Table_nativeConvertColumnToNotNull
                     ThrowException(env, UnsupportedOperation, "The old DateTime type is not supported.");
                     return;
             }
-        }
-        if (table->has_search_index(column_index + 1)) {
-            table->add_search_index(column_index);
         }
         table->remove_column(column_index + 1);
         table->rename_column(table->get_column_index(tmp_column_name), column_name);