+++ realm/realm-library/src/objectServer/java/io/realm/PermissionManager.java
@@ -208,7 +208,7 @@ public class PermissionManager implements Closeable {
         delayedTasks.add(task);
     }
 
-    // Run any tasks that where delayed while the underlying Realms where being opened.
+    // Run any tasks that were delayed while the underlying Realms were being opened.
     // PRECONDITION: Underlying Realms are no longer in the process of being opened.
     private void runDelayedTasks() {
         for (AsyncTask delayedTask : delayedTasks) {
@@ -225,7 +225,8 @@ public class PermissionManager implements Closeable {
     }
 
     // Open both underlying Realms asynchronously. Once they are both ready, all tasks added in the meantime are
-    // started. If the Realms failed to open correctly any error will be reported through the `Callback.onError` callback.
+    // started. Any error will be reported through the `Callback.onError` callback if the Realms failed to open
+    // correctly.
     private void openRealms() {
         if (!openInProgress) {
             openInProgress = true;
@@ -330,7 +331,7 @@ public class PermissionManager implements Closeable {
     }
 
     /**
-     * Checks if this PermissionManager is closed or not. If it is closed, all methods will report back an error
+     * Checks if this PermissionManager is closed or not. If it is closed, all methods will report back an error.
      *
      * @return {@code true} if the PermissionManager is closed, {@code false} if it is still open.
      */
@@ -341,9 +342,8 @@ public class PermissionManager implements Closeable {
 
     @Override
     protected void finalize() throws Throwable {
-        if (managementRealm != null || permissionRealm != null) {
+        if (!closed) {
             RealmLog.warn("PermissionManager was not correctly closed before being finalized.");
-            close();
         }
         super.finalize();
     }
@@ -387,10 +387,8 @@ public class PermissionManager implements Closeable {
                 loadingPermissions.addChangeListener(new RealmChangeListener <RealmResults<Permission>>() {
                     @Override
                     public void onChange(RealmResults <Permission> loadedPermissions) {
-                        // FIXME Until we can figure out why `waitForInitialRemoteData` sometimes
-                        // doesn't work for the permission Realm use the size as a heuristic for
-                        // when the Realm is downloaded. One permission should always be available
-                        // for the Management Realm.
+                        // FIXME Wait until both the __permission and __management Realm are available
+                        // To unblock things we just return whenever either permission is present.
                         if (loadedPermissions.size() > 0) {
                             loadingPermissions.removeChangeListener(this);
                             if (checkAndReportInvalidState()) { return; }