+++ realm/realm-library/src/objectServer/java/io/realm/SyncUser.java
@@ -44,8 +44,8 @@ import io.realm.internal.network.ExponentialBackoffTask;
 import io.realm.internal.network.LogoutResponse;
 import io.realm.internal.objectserver.ObjectServerUser;
 import io.realm.internal.objectserver.Token;
+import io.realm.internal.permissions.ManagementModule;
 import io.realm.log.RealmLog;
-import io.realm.internal.permissions.PermissionModule;
 
 /**
  * This class represents a user on the Realm Object Server. The credentials are provided by various 3rd party
@@ -83,7 +83,7 @@ public class SyncUser {
                                 }
                             }
                         })
-                        .modules(new PermissionModule())
+                        .modules(new ManagementModule())
                         .build();
             }